+++ src/shared/domains.ts
