+++ src/shared/interface.ts
